increase buffer sizes to 8 bytes for 64-bit ints

+++ bones.h
@@ -33,13 +33,13 @@
 
 /*
  * structs from NetHack header files, modified as follows:
- * ints and longs changed to int32 or uint32
+ * ints and longs changed to int64 or uint64
  * xchar, schar, uchar changed to int8 or uint8
  * boolean, aligntyp changed to int8
  * const qualifiers removed
  * bitfields declared without Bitfield() macros
  * void pointers declared without genericptr_t macros
- * pointers changed to uint32 when their values are important
+ * pointers changed to uint64 when their values are important
  * unions replaced with their members (some of which are then unused)
  * crazy mextra and oextra data replaced with regular pointers
  *
@@ -58,14 +58,14 @@ struct version_info {
 #define PL_FSIZ 32 /* fruit name */
 struct fruit {
     char fname[PL_FSIZ];
-    int32 fid;
+    int64 fid;
     struct fruit *nextf;
 };
 
 #define COLNO 80
 #define ROWNO 21
 struct rm {
-    int32 glyph;            /* what the hero thinks is there */
+    int64 glyph;            /* what the hero thinks is there */
     int8 typ;               /* what is really there */
     uint8 seenv;            /* seen vector */
     unsigned flags : 5;       /* extra information for typ */
@@ -138,8 +138,8 @@ struct mkroom {
 
 typedef struct fe {
     struct fe *next;        /* next item in chain */
-    int32 timeout;          /* when we time out */
-    uint32 tid;             /* timer ID */
+    int64 timeout;          /* when we time out */
+    uint64 tid;             /* timer ID */
     int16 kind;             /* kind of use */
     int16 func_index;       /* what to call when we time out */
     void * arg;             /* pointer to timeout argument */
@@ -158,8 +158,8 @@ typedef struct ls_t {
 #define MTSZ    4
 struct monst {
     struct monst *nmon;
-    uint32 data;            /* officially (struct permonst *) */
-    uint32 m_id;
+    uint64 data;            /* officially (struct permonst *) */
+    uint64 m_id;
     int16 mnum;             /* permanent monster index number */
     int16 movement;         /* movement points  */
     uint8 m_lev;            /* adjusted difficulty level of monster */
@@ -167,12 +167,12 @@ struct monst {
     int8 mx, my;
     int8 mux, muy;          /* where the monster thinks you are */
     coord mtrack[MTSZ];     /* monster track */
-    int32 mhp, mhpmax;
-    uint32 mappearance;     /* for undetected mimics and the wiz */
+    int64 mhp, mhpmax;
+    uint64 mappearance;     /* for undetected mimics and the wiz */
     uint8 m_ap_type;        /* what mappearance is describing: */
     int8 mtame;             /* level of tameness, implies peaceful */
     uint16 mintrinsics;     /* low 8 correspond to mresists */
-    int32 mspec_used;       /* monster's special ability attack timeout */
+    int64 mspec_used;       /* monster's special ability attack timeout */
     unsigned female : 1;      /* is female */
     unsigned minvis : 1;      /* currently invisible */
     unsigned invis_blkd : 1;  /* invisibility blocked */
@@ -203,18 +203,18 @@ struct monst {
     unsigned ispriest : 1;    /* is a priest */
     unsigned iswiz : 1;       /* is the Wizard of Yendor */
     unsigned wormno : 5;      /* at most 31 worms on any level */
-    int32 mstrategy;        /* for monsters with mflag3: current strategy */
-    int32 mtrapseen;        /* bitmap of traps we've been trapped in */
-    int32 mlstmv;           /* for catching up with lost time */
-    int32 mgold;            /* ifndef GOLDOBJ */
+    int64 mstrategy;        /* for monsters with mflag3: current strategy */
+    int64 mtrapseen;        /* bitmap of traps we've been trapped in */
+    int64 mlstmv;           /* for catching up with lost time */
+    int64 mgold;            /* ifndef GOLDOBJ */
     struct obj *minvent;
     struct obj *mw;
-    int32 misc_worn_check;
+    int64 misc_worn_check;
     int8 weapon_check;
     uint8 mnamelth;         /* length of name (following mxlth) */
     int16 mxlth;            /* length of following data */
-    int32 meating;          /* monster is eating timeout */
-    int32 mextra[1];        /* monster dependent info */
+    int64 meating;          /* monster is eating timeout */
+    int64 mextra[1];        /* monster dependent info */
 
     /* pointers to mextra data */
     struct egd *egd;
@@ -232,8 +232,8 @@ struct fakecorridor {
 
 #define FCSIZ (ROWNO+COLNO)
 struct egd {
-    int32 fcbeg, fcend;     /* fcend: first unused pos */
-    int32 vroom;            /* room number of the vault */
+    int64 fcbeg, fcend;     /* fcend: first unused pos */
+    int64 vroom;            /* room number of the vault */
     int8 gdx, gdy;          /* goal of guard's walk */
     int8 ogx, ogy;          /* guard's last position */
     d_level gdlevel;        /* level (& dungeon) guard was created in */
@@ -251,20 +251,20 @@ struct epri {
 };
 
 struct bill_x {
-    uint32 bo_id;
+    uint64 bo_id;
     int8 useup;
-    int32 price;            /* price per unit */
-    int32 bquan;            /* amount used up */
+    int64 price;            /* price per unit */
+    int64 bquan;            /* amount used up */
 };
 
 #define BILLSZ 200
 #define PL_NSIZ 32          /* name of player, ghost, shopkeeper */
 struct eshk {
-    int32 robbed;           /* amount stolen by most recent customer */
-    int32 credit;           /* amount credited to customer */
-    int32 debit;            /* amount of debt for using unpaid items */
-    int32 loan;             /* shop-gold picked (part of debit) */
-    int32 shoptype;         /* the value of rooms[shoproom].rtype */
+    int64 robbed;           /* amount stolen by most recent customer */
+    int64 credit;           /* amount credited to customer */
+    int64 debit;            /* amount of debt for using unpaid items */
+    int64 loan;             /* shop-gold picked (part of debit) */
+    int64 shoptype;         /* the value of rooms[shoproom].rtype */
     int8 shoproom;          /* index in rooms; set by inshop() */
     int8 unused;            /* to force alignment for stupid compilers */
     int8 following;         /* following customer since he owes us sth */
@@ -272,10 +272,10 @@ struct eshk {
     coord shk;              /* usual position shopkeeper */
     coord shd;              /* position shop door */
     d_level shoplevel;      /* level (& dungeon) of his shop */
-    int32 billct;           /* no. of entries of bill[] in use */
+    int64 billct;           /* no. of entries of bill[] in use */
     struct bill_x bill[BILLSZ];
     struct bill_x *bill_p;
-    int32 visitct;          /* nr of visits by most recent customer */
+    int64 visitct;          /* nr of visits by most recent customer */
     char customer[PL_NSIZ]; /* most recent customer */
     char shknam[PL_NSIZ];
 };
@@ -285,15 +285,15 @@ struct emin {
 };
 
 struct edog {
-    int32 droptime;         /* moment dog dropped object */
-    uint32 dropdist;        /* dist of drpped obj from @ */
-    int32 apport;           /* amount of training */
-    int32 whistletime;      /* last time he whistled */
-    int32 hungrytime;       /* will get hungry at this time */
+    int64 droptime;         /* moment dog dropped object */
+    uint64 dropdist;        /* dist of drpped obj from @ */
+    int64 apport;           /* amount of training */
+    int64 whistletime;      /* last time he whistled */
+    int64 hungrytime;       /* will get hungry at this time */
     coord ogoal;            /* previous goal location */
-    int32 abuse;            /* track abuses to this pet */
-    int32 revivals;         /* count pet deaths */
-    int32 mhpmax_penalty;   /* while starving, points reduced */
+    int64 abuse;            /* track abuses to this pet */
+    int64 revivals;         /* count pet deaths */
+    int64 mhpmax_penalty;   /* while starving, points reduced */
     unsigned killed_by_u : 1; /* you attempted to kill him */
 };
 
@@ -304,11 +304,11 @@ struct obj {
     struct obj *nobj;
     void *vptr;             /* supposed to be a union of pointers */
     struct obj *cobj;       /* contents list for containers */
-    uint32 o_id;
+    uint64 o_id;
     int8 ox, oy;
     int16 otyp;             /* object class number */
-    uint32 owt;
-    int32 quan;             /* number of items */
+    uint64 owt;
+    int64 quan;             /* number of items */
     int8 spe;               /* meaning depends on type of object */
     int8 oclass;            /* object class */
     int8 invlet;            /* designation in inventory */
@@ -336,17 +336,17 @@ struct obj {
     unsigned oattached : 2;   /* obj struct has special attachment */
     unsigned in_use : 1;      /* ifndef NO_SIGNAL || incarnation>=3.3.0 */
     unsigned bypass : 1;      /* mark this as an object to be skipped by bhito() */
-    int32 corpsenm;         /* type of corpse is mons[corpsenm] */
-    uint32 oeaten;          /* nutrition left in food, if partly eaten */
-    int32 age;              /* creation date */
+    int64 corpsenm;         /* type of corpse is mons[corpsenm] */
+    uint64 oeaten;          /* nutrition left in food, if partly eaten */
+    int64 age;              /* creation date */
     uint8 onamelth;         /* length of name (following oxlth) */
     int16 oxlth;            /* length of following data */
-    int32 owornmask;
-    int32 oextra[1];
+    int64 owornmask;
+    int64 oextra[1];
 
     /* oextra data */
     struct monst *monst;
-    uint32 m_id;
+    uint64 m_id;
     char *oname;
 };
 
@@ -376,14 +376,14 @@ struct engr {
     struct engr *nxt_engr;
     char *engr_txt;
     int8 engr_x, engr_y;
-    uint32 engr_lth;        /* for save & restore; not length of text */
-    int32 engr_time;        /* moment engraving was (will be) finished */
+    uint64 engr_lth;        /* for save & restore; not length of text */
+    int64 engr_time;        /* moment engraving was (will be) finished */
     int8 engr_type;
 };
 
 struct damage {
     struct damage *next;
-    int32 when, cost;
+    int64 when, cost;
     coord place;
     int8 typ;
 };
@@ -398,7 +398,7 @@ typedef struct {
     NhRect *rects;          /* Rectangles composing the region */
     int16 nrects;           /* Number of rectangles  */
     int8 attach_2_u;        /* Region attached to player ? */
-    uint32 attach_2_m;      /* Region attached to monster ? */
+    uint64 attach_2_m;      /* Region attached to monster ? */
     char *enter_msg;        /* Message when entering */
     char *leave_msg;        /* Message when leaving */
     int16 ttl;              /* Time to live. -1 is forever */
@@ -409,12 +409,12 @@ typedef struct {
     int16 leave_f;          /* Function to call when the player leaves */
     int16 inside_f;         /* Function to call every turn player is inside */
     int8 player_flags;
-    uint32 *monsters;       /* Monsters currently inside this region */
+    uint64 *monsters;       /* Monsters currently inside this region */
     int16 n_monst;          /* Number of monsters inside this region */
     int16 max_monst;        /* Maximum number of monsters */
     int8 visible;           /* Is the region visible ? */
-    int32 glyph;            /* Which glyph to use if visible */
-    uint32 arg;             /* Optional user argument */
+    int64 glyph;            /* Which glyph to use if visible */
+    uint64 arg;             /* Optional user argument */
 } NhRegion;
 
 
@@ -426,32 +426,32 @@ typedef struct {
     uint8 nartifacts, monstsz, objsz, flagsz; /* parsed header values */
     char *bonesid;
     struct fruit *fruitchain;
-    int32 hpid;
+    int64 hpid;
     int8 dlvl;
     struct rm locations[COLNO][ROWNO];
-    int32 monstermoves;
+    int64 monstermoves;
     stairway upstair, dnstair, upladder, dnladder, sstairs;
     dest_area updest, dndest;
     struct levelflags flags;
     coord doors[DOORMAX];
-    int32 nroom;
+    int64 nroom;
     struct mkroom *rooms[MAXNROFROOMS+1]; /* the +1 is for a vault */
-    int32 ntimers;
+    int64 ntimers;
     timer_element *timer_base;
-    int32 nlights;
+    int64 nlights;
     light_source *light_base;
-    uint32 monbegin;
+    uint64 monbegin;
     struct monst *monchn;
     struct wseg *wsegments[MAX_NUM_WORMS];
-    int32 wgrowtime[MAX_NUM_WORMS];
+    int64 wgrowtime[MAX_NUM_WORMS];
     struct trap *trapchn;
     struct obj *objchn;
     struct obj *buriedobjs;
     struct obj *billobjs;
     struct engr *engravings;
     struct damage *damage;
-    int32 moves;
-    int32 nregions;
+    int64 moves;
+    int64 nregions;
     NhRegion **regions;
 } bones_t;
